Move MediaWikiClient to a separate file
